+++ apps/innovatech-api/src/db/scripts/seed.ts
@@ -2,17 +2,20 @@ import { AuthService, SignUpCredentialsDto } from '@ivt/a-auth';
 import { UserRoles } from '@ivt/c-data';
 import fs from 'fs';
 import path from 'path';
-import { ConnectionOptions, createConnection } from 'typeorm';
+import { ConnectionOptions, createConnection, getManager } from 'typeorm';
 
 import config from '../config/ormconfig';
 
 async function run() {
   const opt = {
     ...config,
+    synchonize: false,
+    logging: false,
   };
 
   const connection = await createConnection(opt as ConnectionOptions);
   const authService = new AuthService(null, connection);
+  const entityManager = getManager();
 
   const users: SignUpCredentialsDto[] = [
     {
@@ -33,8 +36,6 @@ async function run() {
     },
   ];
 
-  const queryRunner = connection.createQueryRunner();
-
   const filePath = path.join(__dirname, 'sepomex-catalog.sql');
 
   users.forEach(async user => {
@@ -45,24 +46,16 @@ async function run() {
     }
   });
 
-  return new Promise((resolve, reject) => {
-    fs.readFile(filePath, 'utf8', async (error, data: string) => {
-      if (!error) {
-        console.log('Inserting localities...');
-        await queryRunner.query(data);
-        console.log('Inserting localities done!');
-        resolve();
-      } else {
-        console.log(error);
-        reject();
-      }
-    });
-  });
+  console.log('Inserting localities...');
+  try {
+    const data = fs.readFileSync(filePath, 'utf8');
+    await entityManager.query(data);
+    console.log('Inserting localities done!');
+    process.exit(0);
+  } catch (error) {
+    console.log(error);
+    process.exit(0);
+  }
 }
 
-run()
-  .then(() => {
-    console.log('Seeds done');
-    process.exit(0);
-  })
-  .catch(error => console.error('seed error', error));
+run();