+++ libs/api/a-guarantees/src/lib/dto/create-dtos/create-address.dto.ts
@@ -3,7 +3,7 @@ import { IsNotEmpty, IsNumberString, IsOptional, IsString } from 'class-validato
 export class CreateAddressDto {
   @IsNotEmpty()
   @IsNumberString()
-  zipCode: string;
+  zipcode: string;
 
   @IsNotEmpty()
   @IsString()