+++ libs/api/a-products/src/lib/controllers/products.controller.ts
@@ -1,12 +1,25 @@
-import { Body, Controller, Get, Put, Param, ParseIntPipe, Post, Res, UseGuards, Query, ValidationPipe } from '@nestjs/common';
+import { Product } from '@ivt/c-data';
+import {
+  Body,
+  Controller,
+  Get,
+  Param,
+  ParseIntPipe,
+  Post,
+  Put,
+  Query,
+  Res,
+  UseGuards,
+  ValidationPipe,
+} from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+import { Response } from 'express';
 
 import { CreateProductDto } from '../dto/create-products.dto';
-import { UpdateProductDto } from '../dto/update-product.dto';
-import { GetProductsFilterDto} from '../dto/get-products-filter.dto';
 import { GenerateProductPdfDto } from '../dto/generate-product-pdf.dto';
+import { GetProductsFilterDto } from '../dto/get-products-filter.dto';
+import { UpdateProductDto } from '../dto/update-product.dto';
 import { ProductService } from '../services/products.service';
-import { Product } from '@ivt/c-data';
 
 @Controller('products')
 @UseGuards(AuthGuard())
@@ -29,7 +42,10 @@ export class ProductController {
   }
 
   @Post('preview/pdf')
-  async getGuaranteePdf(@Body() generateProductPdfDto: GenerateProductPdfDto, @Res() response: Response): Promise<void> {
+  async getGuaranteePdf(
+    @Body() generateProductPdfDto: GenerateProductPdfDto,
+    @Res() response: Response
+  ): Promise<void> {
     return this.productService.generateProductPdf(generateProductPdfDto, response);
   }