+++ libs/api/a-state/src/lib/guarantees/entities/guarantee.entity.ts
@@ -51,6 +51,7 @@ export class GuaranteeEntity extends BaseEntity implements Guarantee {
       to: value => value,
       from: value => GuaranteeStatus[value],
     },
+    default: GuaranteeStatus.outstanding,
   })
   status: GuaranteeStatus | string;